Rename scope entity.name.function.erlang to meta.function.erlang
This change is motivated by the fact that Erlang code highlighting
looks noticeably different from other languages supported by
Sublime Text. With our Erlang grammar, only punctuation is rendered
using the default text color. Most color schemes are not designed for
that, so Erlang code looks a lot darker.

I have tested this change with all of the color schemes that are bundled
with Sublime Text. It does look better, especially with color
schemes that define a dark background.

Long time users of this package (such as me) might miss the
distinctive coloring of punctuation, but it can easily
be enabled by editing the color scheme file.