More intuitive Goto Definition behaviour for inexact matches
For more than one match, the first match in the current file is pre-selected.
This is what I want in 90% of all cases, especially if the reference is
not module-qualified.

If no exact type or function matches and the reference is qualified,
the module is opened instead (or a panel if the module exists more than once).
This will allow me to scan the export attributes (they're usually
near the module declaration at the top of the file).

If the module is not indexed, a panel listing all definitions of the
symbol is opened even if there is only a single match. This is better
than the previous behaviour (jumping to the first module containing the
symbol, usually one that I didn't want).